requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
PyYAML>=6.0
pydantic>=2.0
tenacity>=8.0
rich>=13.0